requirements.txt
# GPU Management
pynvml==11.5.0
py3nvml==0.2.7
torch>=2.2.0

# Container Management
docker==6.1.3

# HTTP & WebSocket
httpx>=0.23.0,<0.24.0
aiohttp==3.9.0
websockets>=9.0,<12.0

# Blockchain
solana==0.30.2
solders==0.18.0

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0
psutil==5.9.6
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Build tools (optional, for distribution)
pyinstaller==6.3.0
build==1.0.3
twine==4.0.2

